Version 2.6_20190110
Added to Network Summary Report:
	CSDP highest bottom elevation
	CSDP XS within a specified distance of each other
Added Network Coloring, which is coloring of centerlines and
cross-section lines based on various issues, such as -dConveyance, max
area ratio, etc. Network coloring can be toggled on and off using the
Display Parameters menu.
Also in the Display Parameters menu, the Intertidal Low and High tides
can be adjusted. These values are used to warn the user about
-dConveyance issues.

Added a legend panel which describes what these colors mean
In cross-section plot window, -dConvenance values are now displayed in
red.
Changed Multiple cross-section plot: cross-section plots are centered
based on x centroid at a higher elevation (3/4 of the way from the
bottom to the top). This fixes a bug, which was preventing the display
of cross-sections with bottom elevations above zero.
Modified Reach Summary: if you specify a range of channels that includes
a non-existent channel, a dialog will appear asking you to continue or
not.
Network-Calculate now uses an improved dialog, which gives you the
option of creating dsm2 input files in the pre-v8 format, and allows you
to specify a manning's n value that will be used for every channel.