Version 2.6_20190214
Added features for working with centerline representations of levees and
polygons (used to calculate GIS DSM2 channel volumes). Features include:
snap polygon points to levee centerline
Modified export to WKT to include POLYGON vs LINESTRING option
CenterlineSummaryWindow: changed text in data identification labels to
indicate whether the values are for a channel or for a reach.