Version 2.6_20190307
Add average width to network summry report--if a channel is too narrow
relative to the grid size of the DEM used, GIS volume results should not
be used.
Changes to centerline and cross-section editing:
	Points can now be added to either end of a centerline, and to either
end of a cross-section drawing. THe add buttons are replaced with two
add buttons, one for each end of the centerline/cross-section.
	When adding, deleting, and moving centerline points, the CSDP will try
to keep cross-sections in about the same location. (Insert behavior is
not changed, but could be in a later release. Until then, best practice
would be to avoid inserting points too far away from the centerline).
Previously, any change in the length of a centerline upstream from a
cross-section resulted in cross-sections being moved or deleted.
Added Tools-Close Polygon centerlines, which makes sure that the
upstream and downstream ends of a polygon centerline have the same
coordinates.
Added NetworkEditingInfo class, which will be used to store information
about centerline and cross-section line editing to support undo/redo,
which is not yet implemented.
Moved zoom-to-centerline and zoom-to-node from Network Menu to Zoom
Menu.