Version 3.0_20200102
Csdp main class (Csdp) will optionally read a file (path specified as
argument to application), which should contain a string that
will be appended to the version number. This should be the git hash. The
file containing the git hash will not be checked in, but will be
included in the distribution.
Started modifying XsectGraph to not retain unedited dummy cross-sections
(created if no bathymetry data) unless user wishes to retain them.