Version 3.0_20220105
Removed Tools-CreateStraightlineChanForGridmap from Tools menu, because
it is not the way to create a network channel layer. The network channel
layer is instead created by creating new channels for the entire network
(in the network menu). Fixed bug in create DSM2 output locations:
previously, if a landmark was beyond a channel endpoint and user had
specified a channels.inp file to adjust channel lengths to match a
previous version of DSM2, the adjustment was not being applied.